abstract engine mgr
