+++ br/pkg/lightning/backend/local/local.go
@@ -28,7 +28,6 @@ import (
 	"sync"
 	"time"
 
-	"github.com/cockroachdb/pebble"
 	"github.com/coreos/go-semver/semver"
 	"github.com/docker/go-units"
 	"github.com/google/uuid"
@@ -44,14 +43,12 @@ import (
 	"github.com/pingcap/tidb/br/pkg/lightning/config"
 	"github.com/pingcap/tidb/br/pkg/lightning/errormanager"
 	"github.com/pingcap/tidb/br/pkg/lightning/log"
-	"github.com/pingcap/tidb/br/pkg/lightning/manual"
 	"github.com/pingcap/tidb/br/pkg/lightning/metric"
 	"github.com/pingcap/tidb/br/pkg/lightning/tikv"
 	"github.com/pingcap/tidb/br/pkg/logutil"
 	"github.com/pingcap/tidb/br/pkg/membuf"
 	"github.com/pingcap/tidb/br/pkg/pdutil"
 	"github.com/pingcap/tidb/br/pkg/restore/split"
-	"github.com/pingcap/tidb/br/pkg/storage"
 	"github.com/pingcap/tidb/br/pkg/version"
 	"github.com/pingcap/tidb/pkg/infoschema"
 	"github.com/pingcap/tidb/pkg/metrics"
@@ -60,13 +57,10 @@ import (
 	"github.com/pingcap/tidb/pkg/util"
 	"github.com/pingcap/tidb/pkg/util/codec"
 	"github.com/pingcap/tidb/pkg/util/engine"
-	"github.com/tikv/client-go/v2/oracle"
 	tikvclient "github.com/tikv/client-go/v2/tikv"
 	pd "github.com/tikv/pd/client"
 	pdhttp "github.com/tikv/pd/client/http"
-	"go.uber.org/atomic"
 	"go.uber.org/zap"
-	"golang.org/x/sync/errgroup"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/backoff"
 	"google.golang.org/grpc/codes"
@@ -463,9 +457,6 @@ func (c *BackendConfig) adjust() {
 
 // Backend is a local backend.
 type Backend struct {
-	engines        sync.Map // sync version of map[uuid.UUID]*Engine
-	externalEngine map[uuid.UUID]common.Engine
-
 	pdCli            pd.Client
 	pdHTTPCli        pdhttp.Client
 	splitCli         split.SplitClient
@@ -475,13 +466,11 @@ type Backend struct {
 	tikvCodec        tikvclient.Codec
 
 	BackendConfig
+	engineMgr *engineManager
 
 	supportMultiIngest  bool
-	duplicateDB         *pebble.DB
-	keyAdapter          common.KeyAdapter
 	importClientFactory ImportClientFactory
 
-	bufferPool   *membuf.Pool
 	metrics      *metric.Common
 	writeLimiter StoreWriteLimiter
 	logger       log.Logger
@@ -490,28 +479,14 @@ type Backend struct {
 }
 
 var _ DiskUsage = (*Backend)(nil)
+var _ StoreHelper = (*Backend)(nil)
 var _ backend.Backend = (*Backend)(nil)
-
-func openDuplicateDB(storeDir string) (*pebble.DB, error) {
-	dbPath := filepath.Join(storeDir, duplicateDBName)
-	// TODO: Optimize the opts for better write.
-	opts := &pebble.Options{
-		TablePropertyCollectors: []func() pebble.TablePropertyCollector{
-			newRangePropertiesCollector,
-		},
-	}
-	return pebble.Open(dbPath, opts)
-}
 
 const (
 	pdCliMaxMsgSize = int(128 * units.MiB) // pd.ScanRegion may return a large response
 )
 
 var (
-	// RunInTest indicates whether the current process is running in test.
-	RunInTest bool
-	// LastAlloc is the last ID allocator.
-	LastAlloc      manual.Allocator
 	maxCallMsgSize = []grpc.DialOption{
 		grpc.WithDefaultCallOptions(grpc.MaxCallRecvMsgSize(pdCliMaxMsgSize)),
 		grpc.WithDefaultCallOptions(grpc.MaxCallSendMsgSize(pdCliMaxMsgSize)),
@@ -525,12 +500,6 @@ func NewBackend(
 	config BackendConfig,
 	regionSizeGetter TableRegionSizeGetter,
 ) (b *Backend, err error) {
-	var duplicateDB *pebble.DB
-	defer func() {
-		if err != nil && duplicateDB != nil {
-			_ = duplicateDB.Close()
-		}
-	}()
 	config.adjust()
 	pdAddrs := strings.Split(config.PDAddr, ",")
 	pdCli, err := pd.NewClientWithContext(
@@ -563,13 +532,6 @@ func NewBackend(
 		}
 	}
 
-	if config.DupeDetectEnabled {
-		duplicateDB, err = openDuplicateDB(config.LocalStoreDir)
-		if err != nil {
-			return nil, common.ErrOpenDuplicateDB.Wrap(err).GenWithStackByArgs()
-		}
-	}
-
 	// The following copies tikv.NewTxnClient without creating yet another pdClient.
 	spkv, err := tikvclient.NewEtcdSafePointKV(strings.Split(config.PDAddr, ","), tls.TLSConfig())
 	if err != nil {
@@ -595,24 +557,13 @@ func NewBackend(
 	pdHTTPCli := pdhttp.NewClient("lightning", pdAddrs, pdhttp.WithTLSConfig(tls.TLSConfig()))
 	splitCli := split.NewSplitClient(pdCli, pdHTTPCli, tls.TLSConfig(), false)
 	importClientFactory := newImportClientFactoryImpl(splitCli, tls, config.MaxConnPerStore, config.ConnCompressType)
-	keyAdapter := common.KeyAdapter(common.NoopKeyAdapter{})
-	if config.DupeDetectEnabled {
-		keyAdapter = common.DupDetectKeyAdapter{}
-	}
 	var writeLimiter StoreWriteLimiter
 	if config.StoreWriteBWLimit > 0 {
 		writeLimiter = newStoreWriteLimiter(config.StoreWriteBWLimit)
 	} else {
 		writeLimiter = noopStoreWriteLimiter{}
 	}
-	alloc := manual.Allocator{}
-	if RunInTest {
-		alloc.RefCnt = new(atomic.Int64)
-		LastAlloc = alloc
-	}
 	local := &Backend{
-		engines:          sync.Map{},
-		externalEngine:   map[uuid.UUID]common.Engine{},
 		pdCli:            pdCli,
 		pdHTTPCli:        pdHTTPCli,
 		splitCli:         splitCli,
@@ -623,13 +574,15 @@ func NewBackend(
 
 		BackendConfig: config,
 
-		duplicateDB:         duplicateDB,
-		keyAdapter:          keyAdapter,
 		importClientFactory: importClientFactory,
-		bufferPool:          membuf.NewPool(membuf.WithAllocator(alloc)),
 		writeLimiter:        writeLimiter,
 		logger:              log.FromContext(ctx),
 	}
+	engineMgr, err := newEngineManager(config, local, local.logger)
+	if err != nil {
+		return nil, err
+	}
+	local.engineMgr = engineMgr
 	if m, ok := metric.GetCommonMetric(ctx); ok {
 		local.metrics = m
 	}
@@ -642,15 +595,7 @@ func NewBackend(
 
 // TotalMemoryConsume returns the total memory usage of the local backend.
 func (local *Backend) TotalMemoryConsume() int64 {
-	var memConsume int64
-	local.engines.Range(func(k, v interface{}) bool {
-		e := v.(*Engine)
-		if e != nil {
-			memConsume += e.TotalMemorySize()
-		}
-		return true
-	})
-	return memConsume + local.bufferPool.TotalSize()
+	return local.engineMgr.totalMemoryConsume()
 }
 
 func (local *Backend) checkMultiIngestSupport(ctx context.Context) error {
@@ -718,101 +663,12 @@ func (local *Backend) checkMultiIngestSupport(ctx context.Context) error {
 	return nil
 }
 
-// rlock read locks a local file and returns the Engine instance if it exists.
-func (local *Backend) rLockEngine(engineID uuid.UUID) *Engine {
-	if e, ok := local.engines.Load(engineID); ok {
-		engine := e.(*Engine)
-		engine.rLock()
-		return engine
-	}
-	return nil
-}
-
-// lock locks a local file and returns the Engine instance if it exists.
-func (local *Backend) lockEngine(engineID uuid.UUID, state importMutexState) *Engine {
-	if e, ok := local.engines.Load(engineID); ok {
-		engine := e.(*Engine)
-		engine.lock(state)
-		return engine
-	}
-	return nil
-}
-
-// tryRLockAllEngines tries to read lock all engines, return all `Engine`s that are successfully locked.
-func (local *Backend) tryRLockAllEngines() []*Engine {
-	var allEngines []*Engine
-	local.engines.Range(func(k, v interface{}) bool {
-		engine := v.(*Engine)
-		// skip closed engine
-		if engine.tryRLock() {
-			if !engine.closed.Load() {
-				allEngines = append(allEngines, engine)
-			} else {
-				engine.rUnlock()
-			}
-		}
-		return true
-	})
-	return allEngines
-}
-
-// lockAllEnginesUnless tries to lock all engines, unless those which are already locked in the
-// state given by ignoreStateMask. Returns the list of locked engines.
-func (local *Backend) lockAllEnginesUnless(newState, ignoreStateMask importMutexState) []*Engine {
-	var allEngines []*Engine
-	local.engines.Range(func(k, v interface{}) bool {
-		engine := v.(*Engine)
-		if engine.lockUnless(newState, ignoreStateMask) {
-			allEngines = append(allEngines, engine)
-		}
-		return true
-	})
-	return allEngines
-}
-
 // Close the local backend.
 func (local *Backend) Close() {
-	for _, e := range local.externalEngine {
-		_ = e.Close()
-	}
-	local.externalEngine = map[uuid.UUID]common.Engine{}
-	allLocalEngines := local.lockAllEnginesUnless(importMutexStateClose, 0)
-	for _, e := range allLocalEngines {
-		_ = e.Close()
-		e.unlock()
-	}
-	local.engines = sync.Map{}
+	local.engineMgr.close()
 	local.importClientFactory.Close()
-	local.bufferPool.Destroy()
 
-	if local.duplicateDB != nil {
-		// Check if there are duplicates that are not collected.
-		iter := local.duplicateDB.NewIter(&pebble.IterOptions{})
-		hasDuplicates := iter.First()
-		allIsWell := true
-		if err := iter.Error(); err != nil {
-			local.logger.Warn("iterate duplicate db failed", zap.Error(err))
-			allIsWell = false
-		}
-		if err := iter.Close(); err != nil {
-			local.logger.Warn("close duplicate db iter failed", zap.Error(err))
-			allIsWell = false
-		}
-		if err := local.duplicateDB.Close(); err != nil {
-			local.logger.Warn("close duplicate db failed", zap.Error(err))
-			allIsWell = false
-		}
-		// If checkpoint is disabled, or we don't detect any duplicate, then this duplicate
-		// db dir will be useless, so we clean up this dir.
-		if allIsWell && (!local.CheckpointEnabled || !hasDuplicates) {
-			if err := os.RemoveAll(filepath.Join(local.LocalStoreDir, duplicateDBName)); err != nil {
-				local.logger.Warn("remove duplicate db file failed", zap.Error(err))
-			}
-		}
-		local.duplicateDB = nil
-	}
-
-	// if checkpoint is disable or we finish load all data successfully, then files in this
+	// if checkpoint is disabled, or we finish load all data successfully, then files in this
 	// dir will be useless, so we clean up this dir and all files in it.
 	if !local.CheckpointEnabled || common.IsEmptyDir(local.LocalStoreDir) {
 		err := os.RemoveAll(local.LocalStoreDir)
@@ -827,36 +683,12 @@ func (local *Backend) Close() {
 
 // FlushEngine ensure the written data is saved successfully, to make sure no data lose after restart
 func (local *Backend) FlushEngine(ctx context.Context, engineID uuid.UUID) error {
-	engine := local.rLockEngine(engineID)
-
-	// the engine cannot be deleted after while we've acquired the lock identified by UUID.
-	if engine == nil {
-		return errors.Errorf("engine '%s' not found", engineID)
-	}
-	defer engine.rUnlock()
-	if engine.closed.Load() {
-		return nil
-	}
-	return engine.flushEngineWithoutLock(ctx)
+	return local.engineMgr.flushEngine(ctx, engineID)
 }
 
 // FlushAllEngines flush all engines.
 func (local *Backend) FlushAllEngines(parentCtx context.Context) (err error) {
-	allEngines := local.tryRLockAllEngines()
-	defer func() {
-		for _, engine := range allEngines {
-			engine.rUnlock()
-		}
-	}()
-
-	eg, ctx := errgroup.WithContext(parentCtx)
-	for _, engine := range allEngines {
-		e := engine
-		eg.Go(func() error {
-			return e.flushEngineWithoutLock(ctx)
-		})
-	}
-	return eg.Wait()
+	return local.engineMgr.flushAllEngines(parentCtx)
 }
 
 // RetryImportDelay returns the delay time before retrying to import a file.
@@ -869,186 +701,14 @@ func (*Backend) ShouldPostProcess() bool {
 	return true
 }
 
-func (local *Backend) openEngineDB(engineUUID uuid.UUID, readOnly bool) (*pebble.DB, error) {
-	opt := &pebble.Options{
-		MemTableSize: local.MemTableSize,
-		// the default threshold value may cause write stall.
-		MemTableStopWritesThreshold: 8,
-		MaxConcurrentCompactions:    16,
-		// set threshold to half of the max open files to avoid trigger compaction
-		L0CompactionThreshold: math.MaxInt32,
-		L0StopWritesThreshold: math.MaxInt32,
-		LBaseMaxBytes:         16 * units.TiB,
-		MaxOpenFiles:          local.MaxOpenFiles,
-		DisableWAL:            true,
-		ReadOnly:              readOnly,
-		TablePropertyCollectors: []func() pebble.TablePropertyCollector{
-			newRangePropertiesCollector,
-		},
-		DisableAutomaticCompactions: local.DisableAutomaticCompactions,
-	}
-	// set level target file size to avoid pebble auto triggering compaction that split ingest SST files into small SST.
-	opt.Levels = []pebble.LevelOptions{
-		{
-			TargetFileSize: 16 * units.GiB,
-		},
-	}
-
-	dbPath := filepath.Join(local.LocalStoreDir, engineUUID.String())
-	db, err := pebble.Open(dbPath, opt)
-	return db, errors.Trace(err)
-}
-
 // OpenEngine must be called with holding mutex of Engine.
 func (local *Backend) OpenEngine(ctx context.Context, cfg *backend.EngineConfig, engineUUID uuid.UUID) error {
-	db, err := local.openEngineDB(engineUUID, false)
-	if err != nil {
-		return err
-	}
-
-	sstDir := engineSSTDir(local.LocalStoreDir, engineUUID)
-	if !cfg.KeepSortDir {
-		if err := os.RemoveAll(sstDir); err != nil {
-			return errors.Trace(err)
-		}
-	}
-	if !common.IsDirExists(sstDir) {
-		if err := os.Mkdir(sstDir, 0o750); err != nil {
-			return errors.Trace(err)
-		}
-	}
-	engineCtx, cancel := context.WithCancel(ctx)
-
-	e, _ := local.engines.LoadOrStore(engineUUID, &Engine{
-		UUID:               engineUUID,
-		sstDir:             sstDir,
-		sstMetasChan:       make(chan metaOrFlush, 64),
-		ctx:                engineCtx,
-		cancel:             cancel,
-		config:             cfg.Local,
-		tableInfo:          cfg.TableInfo,
-		duplicateDetection: local.DupeDetectEnabled,
-		dupDetectOpt:       local.DuplicateDetectOpt,
-		duplicateDB:        local.duplicateDB,
-		keyAdapter:         local.keyAdapter,
-		logger:             log.FromContext(ctx),
-	})
-	engine := e.(*Engine)
-	engine.lock(importMutexStateOpen)
-	defer engine.unlock()
-	engine.db.Store(db)
-	engine.sstIngester = dbSSTIngester{e: engine}
-	if err = engine.loadEngineMeta(); err != nil {
-		return errors.Trace(err)
-	}
-	if err = local.allocateTSIfNotExists(ctx, engine); err != nil {
-		return errors.Trace(err)
-	}
-	engine.wg.Add(1)
-	go engine.ingestSSTLoop()
-	return nil
-}
-
-func (local *Backend) allocateTSIfNotExists(ctx context.Context, engine *Engine) error {
-	if engine.TS > 0 {
-		return nil
-	}
-	physical, logical, err := local.pdCli.GetTS(ctx)
-	if err != nil {
-		return err
-	}
-	ts := oracle.ComposeTS(physical, logical)
-	engine.TS = ts
-	return engine.saveEngineMeta()
+	return local.engineMgr.openEngine(ctx, cfg, engineUUID)
 }
 
 // CloseEngine closes backend engine by uuid.
 func (local *Backend) CloseEngine(ctx context.Context, cfg *backend.EngineConfig, engineUUID uuid.UUID) error {
-	if externalCfg := cfg.External; externalCfg != nil {
-		storeBackend, err := storage.ParseBackend(externalCfg.StorageURI, nil)
-		if err != nil {
-			return err
-		}
-		store, err := storage.NewWithDefaultOpt(ctx, storeBackend)
-		if err != nil {
-			return err
-		}
-		physical, logical, err := local.pdCli.GetTS(ctx)
-		if err != nil {
-			return err
-		}
-		ts := oracle.ComposeTS(physical, logical)
-		externalEngine := external.NewExternalEngine(
-			store,
-			externalCfg.DataFiles,
-			externalCfg.StatFiles,
-			externalCfg.StartKey,
-			externalCfg.EndKey,
-			externalCfg.SplitKeys,
-			externalCfg.RegionSplitSize,
-			local.keyAdapter,
-			local.DupeDetectEnabled,
-			local.duplicateDB,
-			local.DuplicateDetectOpt,
-			local.WorkerConcurrency,
-			ts,
-			externalCfg.TotalFileSize,
-			externalCfg.TotalKVCount,
-			externalCfg.CheckHotspot,
-		)
-		local.externalEngine[engineUUID] = externalEngine
-		return nil
-	}
-
-	// flush mem table to storage, to free memory,
-	// ask others' advise, looks like unnecessary, but with this we can control memory precisely.
-	engineI, ok := local.engines.Load(engineUUID)
-	if !ok {
-		// recovery mode, we should reopen this engine file
-		db, err := local.openEngineDB(engineUUID, true)
-		if err != nil {
-			return err
-		}
-		engine := &Engine{
-			UUID:               engineUUID,
-			sstMetasChan:       make(chan metaOrFlush),
-			tableInfo:          cfg.TableInfo,
-			keyAdapter:         local.keyAdapter,
-			duplicateDetection: local.DupeDetectEnabled,
-			dupDetectOpt:       local.DuplicateDetectOpt,
-			duplicateDB:        local.duplicateDB,
-			logger:             log.FromContext(ctx),
-		}
-		engine.db.Store(db)
-		engine.sstIngester = dbSSTIngester{e: engine}
-		if err = engine.loadEngineMeta(); err != nil {
-			return err
-		}
-		local.engines.Store(engineUUID, engine)
-		return nil
-	}
-
-	engine := engineI.(*Engine)
-	engine.rLock()
-	if engine.closed.Load() {
-		engine.rUnlock()
-		return nil
-	}
-
-	err := engine.flushEngineWithoutLock(ctx)
-	engine.rUnlock()
-
-	// use mutex to make sure we won't close sstMetasChan while other routines
-	// trying to do flush.
-	engine.lock(importMutexStateClose)
-	engine.closed.Store(true)
-	close(engine.sstMetasChan)
-	engine.unlock()
-	if err != nil {
-		return errors.Trace(err)
-	}
-	engine.wg.Wait()
-	return engine.ingestErr.Get()
+	return local.engineMgr.closeEngine(ctx, cfg, engineUUID)
 }
 
 func (local *Backend) getImportClient(ctx context.Context, storeID uint64) (sst.ImportSSTClient, error) {
@@ -1518,10 +1178,10 @@ func (local *Backend) ImportEngine(
 	regionSplitSize, regionSplitKeys int64,
 ) error {
 	var e common.Engine
-	if externalEngine, ok := local.externalEngine[engineUUID]; ok {
+	if externalEngine, ok := local.engineMgr.getExternalEngine(engineUUID); ok {
 		e = externalEngine
 	} else {
-		localEngine := local.lockEngine(engineUUID, importMutexStateImport)
+		localEngine := local.engineMgr.lockEngine(engineUUID, importMutexStateImport)
 		if localEngine == nil {
 			// skip if engine not exist. See the comment of `CloseEngine` for more detail.
 			return nil
@@ -1750,101 +1410,23 @@ func (local *Backend) doImport(ctx context.Context, engine common.Engine, region
 
 // GetImportedKVCount returns the number of imported KV pairs of some engine.
 func (local *Backend) GetImportedKVCount(engineUUID uuid.UUID) int64 {
-	v, ok := local.engines.Load(engineUUID)
-	if !ok {
-		// we get it after import, but before clean up, so this should not happen
-		// todo: return error
-		return 0
-	}
-	e := v.(*Engine)
-	return e.importedKVCount.Load()
+	return local.engineMgr.getImportedKVCount(engineUUID)
 }
 
 // GetExternalEngineKVStatistics returns kv statistics of some engine.
 func (local *Backend) GetExternalEngineKVStatistics(engineUUID uuid.UUID) (
 	totalKVSize int64, totalKVCount int64) {
-	v, ok := local.externalEngine[engineUUID]
-	if !ok {
-		return 0, 0
-	}
-	return v.ImportedStatistics()
+	return local.engineMgr.getExternalEngineKVStatistics(engineUUID)
 }
 
 // ResetEngine reset the engine and reclaim the space.
 func (local *Backend) ResetEngine(ctx context.Context, engineUUID uuid.UUID) error {
-	// the only way to reset the engine + reclaim the space is to delete and reopen it 🤷
-	localEngine := local.lockEngine(engineUUID, importMutexStateClose)
-	if localEngine == nil {
-		if engineI, ok := local.externalEngine[engineUUID]; ok {
-			extEngine := engineI.(*external.Engine)
-			return extEngine.Reset()
-		}
-
-		log.FromContext(ctx).Warn("could not find engine in cleanupEngine", zap.Stringer("uuid", engineUUID))
-		return nil
-	}
-	defer localEngine.unlock()
-	if err := localEngine.Close(); err != nil {
-		return err
-	}
-	if err := localEngine.Cleanup(local.LocalStoreDir); err != nil {
-		return err
-	}
-	db, err := local.openEngineDB(engineUUID, false)
-	if err == nil {
-		localEngine.db.Store(db)
-		localEngine.engineMeta = engineMeta{}
-		if !common.IsDirExists(localEngine.sstDir) {
-			if err := os.Mkdir(localEngine.sstDir, 0o750); err != nil {
-				return errors.Trace(err)
-			}
-		}
-		if err = local.allocateTSIfNotExists(ctx, localEngine); err != nil {
-			return errors.Trace(err)
-		}
-		failpoint.Inject("mockAllocateTSErr", func() {
-			// mock generate timestamp error when reset engine.
-			localEngine.TS = 0
-			mockGRPCErr, _ := status.FromError(errors.Errorf("mock generate timestamp error"))
-			failpoint.Return(errors.Trace(mockGRPCErr.Err()))
-		})
-	}
-	localEngine.pendingFileSize.Store(0)
-
-	return err
+	return local.engineMgr.resetEngine(ctx, engineUUID)
 }
 
 // CleanupEngine cleanup the engine and reclaim the space.
 func (local *Backend) CleanupEngine(ctx context.Context, engineUUID uuid.UUID) error {
-	localEngine := local.lockEngine(engineUUID, importMutexStateClose)
-	// release this engine after import success
-	if localEngine == nil {
-		if extEngine, ok := local.externalEngine[engineUUID]; ok {
-			retErr := extEngine.Close()
-			delete(local.externalEngine, engineUUID)
-			return retErr
-		}
-		log.FromContext(ctx).Warn("could not find engine in cleanupEngine", zap.Stringer("uuid", engineUUID))
-		return nil
-	}
-	defer localEngine.unlock()
-
-	// since closing the engine causes all subsequent operations on it panic,
-	// we make sure to delete it from the engine map before calling Close().
-	// (note that Close() returning error does _not_ mean the pebble DB
-	// remains open/usable.)
-	local.engines.Delete(engineUUID)
-	err := localEngine.Close()
-	if err != nil {
-		return err
-	}
-	err = localEngine.Cleanup(local.LocalStoreDir)
-	if err != nil {
-		return err
-	}
-	localEngine.TotalSize.Store(0)
-	localEngine.Length.Store(0)
-	return nil
+	return local.engineMgr.cleanupEngine(ctx, engineUUID)
 }
 
 // GetDupeController returns a new dupe controller.
@@ -1855,8 +1437,8 @@ func (local *Backend) GetDupeController(dupeConcurrency int, errorMgr *errormana
 		tikvCodec:           local.tikvCodec,
 		errorMgr:            errorMgr,
 		dupeConcurrency:     dupeConcurrency,
-		duplicateDB:         local.duplicateDB,
-		keyAdapter:          local.keyAdapter,
+		duplicateDB:         local.engineMgr.getDuplicateDB(),
+		keyAdapter:          local.engineMgr.getKeyAdapter(),
 		importClientFactory: local.importClientFactory,
 		resourceGroupName:   local.ResourceGroupName,
 		taskType:            local.TaskType,
@@ -1878,7 +1460,7 @@ func (local *Backend) UnsafeImportAndReset(ctx context.Context, engineUUID uuid.
 	if err := closedEngine.Import(ctx, regionSplitSize, regionSplitKeys); err != nil {
 		return err
 	}
-	return local.ResetEngine(ctx, engineUUID)
+	return local.engineMgr.resetEngine(ctx, engineUUID)
 }
 
 func engineSSTDir(storeDir string, engineUUID uuid.UUID) string {
@@ -1886,13 +1468,8 @@ func engineSSTDir(storeDir string, engineUUID uuid.UUID) string {
 }
 
 // LocalWriter returns a new local writer.
-func (local *Backend) LocalWriter(_ context.Context, cfg *backend.LocalWriterConfig, engineUUID uuid.UUID) (backend.EngineWriter, error) {
-	e, ok := local.engines.Load(engineUUID)
-	if !ok {
-		return nil, errors.Errorf("could not find engine for %s", engineUUID.String())
-	}
-	engine := e.(*Engine)
-	return openLocalWriter(cfg, engine, local.tikvCodec, local.LocalWriterMemCacheSize, local.bufferPool.NewBuffer())
+func (local *Backend) LocalWriter(ctx context.Context, cfg *backend.LocalWriterConfig, engineUUID uuid.UUID) (backend.EngineWriter, error) {
+	return local.engineMgr.localWriter(ctx, cfg, engineUUID)
 }
 
 // SwitchModeByKeyRanges will switch tikv mode for regions in the specific key range for multirocksdb.
@@ -1987,12 +1564,17 @@ func nextKey(key []byte) []byte {
 
 // EngineFileSizes implements DiskUsage interface.
 func (local *Backend) EngineFileSizes() (res []backend.EngineFileSize) {
-	local.engines.Range(func(k, v interface{}) bool {
-		engine := v.(*Engine)
-		res = append(res, engine.getEngineFileSize())
-		return true
-	})
-	return
+	return local.engineMgr.engineFileSizes()
+}
+
+// GetTSO implements StoreHelper interface.
+func (local *Backend) GetTSO(ctx context.Context) (physical, logical int64, err error) {
+	return local.pdCli.GetTS(ctx)
+}
+
+// GetTiKVCodec implements StoreHelper interface.
+func (local *Backend) GetTiKVCodec() tikvclient.Codec {
+	return local.tikvCodec
 }
 
 var getSplitConfFromStoreFunc = getSplitConfFromStore
